requirements.txt
# Flask Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5

# AI Integration
openai>=1.0.0

# Environment Management
python-dotenv==1.0.0

# Security
cryptography==41.0.4
bcrypt==4.0.1
Werkzeug==2.3.7

# Development Tools
gunicorn==21.2.0

# Optional: Alternative AI providers
# anthropic==0.3.11
# google-generativeai==0.1.0
